+++ backends/stablehlo/stablehlo_test.go
@@ -1,4 +1,4 @@
-package simplego
+package stablehlo_test
 
 import (
 	"fmt"
@@ -6,7 +6,10 @@ import (
 	"testing"
 
 	"github.com/gomlx/gomlx/backends"
+	"github.com/gomlx/gomlx/backends/stablehlo"
+	"github.com/gomlx/gomlx/graph"
 	"github.com/janpfeifer/must"
+	"github.com/stretchr/testify/assert"
 	"k8s.io/klog/v2"
 )
 
@@ -18,14 +21,14 @@ func init() {
 
 func setup() {
 	fmt.Printf("Available backends: %q\n", backends.List())
-	// Perform your setup logic here
 	if os.Getenv(backends.ConfigEnvVar) == "" {
-		must.M(os.Setenv(backends.ConfigEnvVar, "go"))
+		must.M(os.Setenv(backends.ConfigEnvVar, stablehlo.BackendName))
 	} else {
 		fmt.Printf("\t$%s=%q\n", backends.ConfigEnvVar, os.Getenv(backends.ConfigEnvVar))
 	}
 	backend = backends.MustNew()
 	fmt.Printf("Backend: %s, %s\n", backend.Name(), backend.Description())
+	fmt.Printf("\t- Add flag -vmodule=executable=2 to log the StableHLO program being executed.\n")
 }
 
 func teardown() {
@@ -37,4 +40,11 @@ func TestMain(m *testing.M) {
 	code := m.Run() // Run all tests in the file
 	teardown()
 	os.Exit(code)
+}
+
+func TestCompileAndRun(t *testing.T) {
+	// Just return a constant.
+	exec := graph.NewExec(backend, func(g *graph.Graph) *graph.Node { return graph.Const(g, float32(-7)) })
+	y0 := exec.Call()[0]
+	assert.Equal(t, float32(-7), y0.Value())
 }